Split k0smotroncluster controller in several files
