+++ internal/controller/k0smotroncluster_controller.go
@@ -21,21 +21,14 @@ import (
 	"fmt"
 	"time"
 
-	apps "k8s.io/api/apps/v1"
-	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/util/intstr"
-	informerv1 "k8s.io/client-go/informers/core/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/cache"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
 	km "github.com/k0sproject/k0smotron/api/v1beta1"
-	"github.com/k0sproject/k0smotron/internal/exec"
 )
 
 const (
@@ -125,328 +118,9 @@ func (r *K0smotronClusterReconciler) updateStatus(ctx context.Context, kmc km.K0
 	}
 }
 
-func (r *K0smotronClusterReconciler) reconcileCM(ctx context.Context, req ctrl.Request, kmc km.K0smotronCluster) error {
-	logger := log.FromContext(ctx)
-	logger.Info("Reconciling configmap")
-
-	if kmc.Spec.Service.Type == v1.ServiceTypeLoadBalancer && kmc.Spec.ExternalAddress == "" {
-		return nil
-	}
-
-	cm := r.generateCM(&kmc)
-
-	return r.Client.Patch(ctx, &cm, client.Apply, patchOpts...)
-}
-
-func (r *K0smotronClusterReconciler) reconcileServices(ctx context.Context, req ctrl.Request, kmc km.K0smotronCluster) error {
-	logger := log.FromContext(ctx)
-	// Depending on ingress configuration create nodePort service.
-	logger.Info("Reconciling services")
-	svc := r.generateService(&kmc)
-	if kmc.Spec.Service.Type == v1.ServiceTypeLoadBalancer && kmc.Spec.ExternalAddress == "" {
-		err := r.watchLBServiceIP(ctx, kmc)
-		if err != nil {
-			return err
-		}
-	}
-	return r.Client.Patch(ctx, &svc, client.Apply, patchOpts...)
-}
-
-func (r *K0smotronClusterReconciler) reconcileStatefulSet(ctx context.Context, req ctrl.Request, kmc km.K0smotronCluster) error {
-	logger := log.FromContext(ctx)
-	logger.Info("Reconciling statefulset")
-	statefulSet := r.generateStatefulSet(&kmc)
-
-	return r.Client.Patch(ctx, &statefulSet, client.Apply, patchOpts...)
-}
-
 // SetupWithManager sets up the controller with the Manager.
 func (r *K0smotronClusterReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
 		For(&km.K0smotronCluster{}).
 		Complete(r)
-}
-
-func (r *K0smotronClusterReconciler) generateService(kmc *km.K0smotronCluster) v1.Service {
-	var name string
-	switch kmc.Spec.Service.Type {
-	case v1.ServiceTypeNodePort:
-		name = kmc.GetNodePortName()
-	case v1.ServiceTypeLoadBalancer:
-		name = kmc.GetLoadBalancerName()
-	}
-
-	svc := v1.Service{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: "v1",
-			Kind:       "Service",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: kmc.Namespace,
-			Labels:    map[string]string{"app": "k0smotron"},
-		},
-		Spec: v1.ServiceSpec{
-			Type:     kmc.Spec.Service.Type,
-			Selector: map[string]string{"app": "k0smotron"},
-			Ports: []v1.ServicePort{
-				{
-					Port:       int32(kmc.Spec.Service.APIPort),
-					TargetPort: intstr.FromInt(kmc.Spec.Service.APIPort),
-					Name:       "api",
-					NodePort:   int32(kmc.Spec.Service.APIPort),
-				},
-				{
-					Port:       int32(kmc.Spec.Service.KonnectivityPort),
-					TargetPort: intstr.FromInt(kmc.Spec.Service.KonnectivityPort),
-					Name:       "konnectivity",
-					NodePort:   int32(kmc.Spec.Service.KonnectivityPort),
-				},
-			},
-		},
-	}
-
-	_ = ctrl.SetControllerReference(kmc, &svc, r.Scheme)
-
-	return svc
-}
-
-func (r *K0smotronClusterReconciler) generateCM(kmc *km.K0smotronCluster) v1.ConfigMap {
-	// TODO externalAddress cannot be hardcoded
-	// TODO k0s.yaml should probably be a
-	// github.com/k0sproject/k0s/pkg/apis/k0s.k0sproject.io/v1beta1.ClusterConfig
-	// and then unmarshalled into json to make modification of fields reliable
-	cm := v1.ConfigMap{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "ConfigMap",
-			APIVersion: "v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      kmc.GetConfigMapName(),
-			Namespace: kmc.Namespace,
-		},
-		Data: map[string]string{
-			"k0s.yaml": fmt.Sprintf(`apiVersion: k0s.k0sproject.io/v1beta1
-kind: ClusterConfig
-metadata:
-  name: k0s
-spec:
-  api:
-    externalAddress: %s
-    port: %d
-  konnectivity:
-    agentPort: %d`, kmc.Spec.ExternalAddress, kmc.Spec.Service.APIPort, kmc.Spec.Service.KonnectivityPort), // TODO: do it as a template or something like this
-		},
-	}
-
-	_ = ctrl.SetControllerReference(kmc, &cm, r.Scheme)
-	return cm
-}
-
-func (r *K0smotronClusterReconciler) generateStatefulSet(kmc *km.K0smotronCluster) apps.StatefulSet {
-	k0sVersion := kmc.Spec.K0sVersion
-	if k0sVersion == "" {
-		k0sVersion = defaultK0SVersion
-	}
-
-	statefulSet := apps.StatefulSet{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: "apps/v1",
-			Kind:       "StatefulSet",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      kmc.GetStatefulSetName(),
-			Namespace: kmc.Namespace,
-			Labels:    map[string]string{"app": "k0smotron"},
-		},
-		Spec: apps.StatefulSetSpec{
-			Selector: &metav1.LabelSelector{
-				MatchLabels: map[string]string{"app": "k0smotron"},
-			},
-			Template: v1.PodTemplateSpec{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: map[string]string{"app": "k0smotron"},
-				},
-				Spec: v1.PodSpec{
-					Containers: []v1.Container{{
-						Name:            "controller",
-						Image:           fmt.Sprintf("%s:%s", kmc.Spec.K0sImage, k0sVersion),
-						ImagePullPolicy: v1.PullIfNotPresent,
-						Args:            []string{"k0s", "controller", "--config", "/etc/k0s/k0s.yaml"},
-						Ports: []v1.ContainerPort{
-							{
-								Name:          "api",
-								Protocol:      v1.ProtocolTCP,
-								ContainerPort: int32(kmc.Spec.Service.APIPort),
-							},
-							{
-								Name:          "konnectivity",
-								Protocol:      v1.ProtocolTCP,
-								ContainerPort: int32(kmc.Spec.Service.KonnectivityPort),
-							},
-						},
-						VolumeMounts: []v1.VolumeMount{{
-							Name:      "k0s-config",
-							MountPath: "/etc/k0s",
-							ReadOnly:  true,
-						}},
-						ReadinessProbe: &v1.Probe{ProbeHandler: v1.ProbeHandler{Exec: &v1.ExecAction{Command: []string{"k0s", "status"}}}},
-						LivenessProbe:  &v1.Probe{ProbeHandler: v1.ProbeHandler{Exec: &v1.ExecAction{Command: []string{"k0s", "status"}}}},
-					}},
-					Volumes: []v1.Volume{{
-						Name: "k0s-config",
-						VolumeSource: v1.VolumeSource{
-							// TODO LocalObjectReference can't be hardcoded
-							ConfigMap: &v1.ConfigMapVolumeSource{
-								LocalObjectReference: v1.LocalObjectReference{Name: kmc.GetConfigMapName()},
-								Items: []v1.KeyToPath{{
-									Key:  "k0s.yaml",
-									Path: "k0s.yaml",
-								}},
-							}}}},
-				}},
-		}}
-
-	switch kmc.Spec.Persistence.Type {
-	case "emptyDir":
-		statefulSet.Spec.Template.Spec.Volumes = append(statefulSet.Spec.Template.Spec.Volumes, v1.Volume{
-			Name: kmc.GetVolumeName(),
-			VolumeSource: v1.VolumeSource{
-				EmptyDir: &v1.EmptyDirVolumeSource{},
-			},
-		})
-		statefulSet.Spec.Template.Spec.Containers[0].VolumeMounts = append(statefulSet.Spec.Template.Spec.Containers[0].VolumeMounts, v1.VolumeMount{
-			Name:      kmc.GetVolumeName(),
-			MountPath: "/var/lib/k0s",
-		})
-	case "hostPath":
-		statefulSet.Spec.Template.Spec.Volumes = append(statefulSet.Spec.Template.Spec.Volumes, v1.Volume{
-			Name: kmc.GetVolumeName(),
-			VolumeSource: v1.VolumeSource{
-				HostPath: &v1.HostPathVolumeSource{
-					Path: kmc.Spec.Persistence.HostPath,
-				},
-			},
-		})
-		statefulSet.Spec.Template.Spec.Containers[0].VolumeMounts = append(statefulSet.Spec.Template.Spec.Containers[0].VolumeMounts, v1.VolumeMount{
-			Name:      kmc.GetVolumeName(),
-			MountPath: "/var/lib/k0s",
-		})
-	case "pvc":
-		statefulSet.Spec.VolumeClaimTemplates = append(statefulSet.Spec.VolumeClaimTemplates, kmc.Spec.Persistence.PersistentVolumeClaim)
-
-		statefulSet.Spec.Template.Spec.Containers[0].VolumeMounts = append(statefulSet.Spec.Template.Spec.Containers[0].VolumeMounts, v1.VolumeMount{
-			Name:      kmc.GetVolumeName(),
-			MountPath: "/var/lib/k0s",
-		})
-	}
-
-	ctrl.SetControllerReference(kmc, &statefulSet, r.Scheme)
-	return statefulSet
-}
-
-func (r *K0smotronClusterReconciler) reconcileKubeConfigSecret(ctx context.Context, kmc km.K0smotronCluster) error {
-	logger := log.FromContext(ctx)
-	pod, err := r.findStatefulSetPod(ctx, kmc.GetStatefulSetName(), kmc.Namespace)
-
-	if err != nil {
-		return err
-	}
-
-	output, err := exec.PodExecCmdOutput(ctx, r.ClientSet, r.RESTConfig, pod.Name, kmc.Namespace, "k0s kubeconfig admin")
-	if err != nil {
-		return err
-	}
-
-	logger.Info("Kubeconfig generated, creating the secret")
-
-	secret := v1.Secret{
-		// The dynamic r.Client needs TypeMeta to be set
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Secret",
-			APIVersion: "v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      kmc.GetAdminConfigSecretName(),
-			Namespace: kmc.Namespace,
-			Labels:    map[string]string{"app": "k0smotron"},
-		},
-		StringData: map[string]string{"kubeconfig": output},
-	}
-
-	if err = ctrl.SetControllerReference(&kmc, &secret, r.Scheme); err != nil {
-		return err
-	}
-
-	return r.Client.Patch(ctx, &secret, client.Apply, patchOpts...)
-}
-
-// findStatefulSetPod returns a first running pod from a StatefulSet
-func (r *K0smotronClusterReconciler) findStatefulSetPod(ctx context.Context, statefulSet string, namespace string) (*v1.Pod, error) {
-	dep, err := r.ClientSet.AppsV1().StatefulSets(namespace).Get(ctx, statefulSet, metav1.GetOptions{})
-	if err != nil {
-		return nil, err
-	}
-	selector := metav1.FormatLabelSelector(dep.Spec.Selector)
-	pods, err := r.ClientSet.CoreV1().Pods(namespace).List(ctx, metav1.ListOptions{
-		LabelSelector: selector,
-	})
-	if err != nil {
-		return nil, err
-	}
-	if len(pods.Items) < 1 {
-		return nil, fmt.Errorf("did not find matching pods for statefulSet %s", statefulSet)
-	}
-	// Find a running pod
-	var runningPod *v1.Pod
-	for _, p := range pods.Items {
-		if p.Status.Phase == v1.PodRunning {
-			runningPod = &p
-			break
-		}
-	}
-	if runningPod == nil {
-		return nil, fmt.Errorf("did not find running pods for statefulSet %s", statefulSet)
-	}
-	return runningPod, nil
-}
-
-func (r *K0smotronClusterReconciler) watchLBServiceIP(ctx context.Context, kmc km.K0smotronCluster) error {
-	var handler cache.ResourceEventHandlerFuncs
-	stopCh := make(chan struct{})
-	handler.UpdateFunc = func(old, new interface{}) {
-		newObj, ok := new.(*v1.Service)
-		if !ok {
-			return
-		}
-
-		var externalAddress string
-		if len(newObj.Status.LoadBalancer.Ingress) > 0 {
-			if newObj.Status.LoadBalancer.Ingress[0].Hostname != "" {
-				externalAddress = newObj.Status.LoadBalancer.Ingress[0].Hostname
-			}
-			if newObj.Status.LoadBalancer.Ingress[0].IP != "" {
-				externalAddress = newObj.Status.LoadBalancer.Ingress[0].IP
-			}
-
-			if externalAddress != "" {
-				kmc.Spec.ExternalAddress = externalAddress
-				err := r.Client.Update(ctx, &kmc)
-				if err != nil {
-					log.FromContext(ctx).Error(err, "failed to update K0smotronCluster")
-				}
-				stopCh <- struct{}{}
-			}
-		}
-	}
-
-	inf := informerv1.NewServiceInformer(r.ClientSet, kmc.Namespace, 100*time.Millisecond, nil)
-	_, err := inf.AddEventHandler(handler)
-	if err != nil {
-		return err
-	}
-
-	go inf.Run(stopCh)
-
-	return nil
 }
